requirements.txt
google-cloud-aiplatform>=1.36.0
google-cloud-vision>=3.5.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=0.8.11
tika>=2.6.0
numpy>=1.24.0
langchain>=0.1.0
pytest>=7.4.0
google-generativeai>=0.3.2
aiohttp>=3.9.0
pyyaml>=6.0.0
motor>=3.3.0 